+++ client/src/__tests__/LeaderBoard.test.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { render, screen } from "@testing-library/react"
-import LeaderBoard from "../components/CreateGame/Lecturer/Temp/Temp"
+import LeaderBoard from "../components/CreateGame/Lecturer/Leaderboard/Leaderboard"
 
 describe("LeaderBoard component", () => {
     const teamName = "Your Team"
